move flowChart in domain overview
